This looks good

+++ Patches/ItemDropPatches.cs
@@ -5,7 +5,7 @@ namespace VMP_Mod.Patches
 {
     class ItemDropPatches
     {
-        public static ConfigEntry<int> WeightReduction;
+        public static ConfigEntry<float> WeightReduction;
         public static ConfigEntry<int> itemStackMultiplier;
         public static ConfigEntry<bool> NoTeleportPrevention;
 
@@ -21,7 +21,7 @@ namespace VMP_Mod.Patches
 
                 if (itemStackMultiplier.Value > 0)
                 {
-
+                    //how much do we use this fr?
                     __instance.m_itemData.m_shared.m_weight /= WeightReduction.Value;
 
                     if (__instance.m_itemData.m_shared.m_maxStackSize > 1)
